subject: Appointment section made responsive
body: it's iniciated inside Home Component that been made responsive.
