+++ src/pages/Home/Home.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Appointment from '../../components/appointment/Appointment'
+import Appointment from '../../components/Appointment/Appointment'
 
 function Home() {
   return (<div className="Home">
